refactor: DRY up Title Reveal sound/MIDI code
- Extract shared playSound to lib/utils/audioPlayer.ts (was duplicated
  in TitleRevealRenderer and WordHarvestRenderer)
- Move AUDIO_TYPES to lib/filetypes.ts alongside IMAGE_TYPES/VIDEO_TYPES
- Reuse TitleRevealDefaults type and TitleRevealSaveData in panel
- Derive DEFAULT_TITLE_REVEAL_DEFAULTS from Zod schema
- Add consistent MIDI input clamping in editor
- Add "title-sounds" to UploadSubfolder union type
